Issue #1664: Move ToolbarViewModel to navigationoverlay pkg.
The toolbar and navigationoverlay packages both had one item in them.
The toolbar is only used in the navigationoverlay so I didn't see any
reason to keep around two very small packages.
